Create randomizer_javascript
This JavaScript image randomizer takes an array of filenames, randomly chooses one of them, and then displays that image on a webpage. The randomizer is active whenever a page is loaded. Images do not change once a page is loaded but a timer can be added to activate the randomizer while a page is already loaded.